refactor(userlist): move UserListItem and UserListsScreen to component package
